+++ api/lib/search/agol.ts
@@ -7,8 +7,8 @@ import { Feature } from '@tak-ps/node-cot';
 import { CoTParser } from '@tak-ps/node-cot';
 import ArcGISTokenManager from './arcgis-token-manager.js';
 import ArcGISConfigService from './arcgis-config.js'
-import { FetchSuggest, FetchReverse, FetchForward } from '../search.js'
-import { Search } from '../search.js'
+import { Search } from '../search.js';
+import { FetchSuggest, FetchReverse, FetchForward } from './types.js'
 
 export const AGOLReverseContainer = Type.Object({
     address: Type.Optional(FetchReverse),
