+++ api/routes/search.ts
@@ -4,7 +4,8 @@ import Schema from '@openaddresses/batch-schema';
 import Err from '@openaddresses/batch-error';
 import Auth from '../lib/auth.js';
 import Weather, { FetchHourly } from '../lib/weather.js';
-import { SearchManager, SearchManagerConfig, FetchReverse, FetchSuggest, FetchForward } from '../lib/search.js';
+import { SearchManager } from '../lib/search.js';
+import { SearchManagerConfig, FetchReverse, FetchSuggest, FetchForward } from '../lib/search/types.js';
 import { Feature } from '@tak-ps/node-cot';
 import Config from '../lib/config.js';
